pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "LeadHit test case"
authors = ["Vasilii Stakrotckii <142693897+inferno681@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
motor = "^3.6.0"
uvicorn = "^0.32.1"
pydantic-settings = "^2.6.1"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
wemake-python-styleguide = "^0.19.2"
isort = "^5.13.2"
mypy = "^1.13.0"
types-pyyaml = "^6.0.12.20240917"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
